+++ apps/backend/src/modules/user-follows/user-follows.service.test.ts
@@ -83,7 +83,7 @@ describe("UserFollowsService", () => {
 					false,
 				)) as UserFollowsRepository["checkIsUserFollowing"],
 			followUser: (() =>
-				Promise.resolve()) as UserFollowsRepository["followUser"],
+				Promise.resolve(true)) as UserFollowsRepository["followUser"],
 			unfollowUser: (() =>
 				Promise.resolve(
 					Boolean(DELETED_COUNT_ONE),
@@ -114,7 +114,7 @@ describe("UserFollowsService", () => {
 			checkIsUserFollowing: (() =>
 				Promise.resolve(true)) as UserFollowsRepository["checkIsUserFollowing"],
 			followUser: (() =>
-				Promise.resolve()) as UserFollowsRepository["followUser"],
+				Promise.resolve(false)) as UserFollowsRepository["followUser"],
 			unfollowUser: (() =>
 				Promise.resolve(
 					Boolean(DELETED_COUNT_ONE),
@@ -147,7 +147,7 @@ describe("UserFollowsService", () => {
 					false,
 				)) as UserFollowsRepository["checkIsUserFollowing"],
 			followUser: (() =>
-				Promise.resolve()) as UserFollowsRepository["followUser"],
+				Promise.resolve(true)) as UserFollowsRepository["followUser"],
 			unfollowUser: (() =>
 				Promise.resolve(
 					Boolean(DELETED_COUNT_ONE),
@@ -170,7 +170,7 @@ describe("UserFollowsService", () => {
 					false,
 				)) as UserFollowsRepository["checkIsUserFollowing"],
 			followUser: (() =>
-				Promise.resolve()) as UserFollowsRepository["followUser"],
+				Promise.resolve(false)) as UserFollowsRepository["followUser"],
 			unfollowUser: (() =>
 				Promise.resolve(
 					Boolean(DELETED_COUNT_ONE),
@@ -203,7 +203,7 @@ describe("UserFollowsService", () => {
 					false,
 				)) as UserFollowsRepository["checkIsUserFollowing"],
 			followUser: (() =>
-				Promise.resolve()) as UserFollowsRepository["followUser"],
+				Promise.resolve(false)) as UserFollowsRepository["followUser"],
 			unfollowUser: (() =>
 				Promise.resolve(
 					Boolean(DELETED_COUNT_NONE),
@@ -231,7 +231,7 @@ describe("UserFollowsService", () => {
 			checkIsUserFollowing: (() =>
 				Promise.resolve(true)) as UserFollowsRepository["checkIsUserFollowing"],
 			followUser: (() =>
-				Promise.resolve()) as UserFollowsRepository["followUser"],
+				Promise.resolve(true)) as UserFollowsRepository["followUser"],
 			unfollowUser: (() =>
 				Promise.resolve(
 					Boolean(DELETED_COUNT_ONE),