+++ apps/frontend/src/libs/components/select/libs/types/types.ts
@@ -1,1 +1,0 @@
-export { type SelectOption } from "./select-option.type.js";